requirements.txt
# pip install -r requirements.txt

# base ----------------------------------------

pandas~=1.2.5
pyqt~=5.9.2
pandas~=1.2.5